# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabinet CRM — façade de services",
    "description": "Façade HTTP localhost réutilisant le moteur Python du CRM. Toutes les routes /api/* exigent l'en-tête Authorization: Bearer <jeton>.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/jobs": {
      "get": {
        "tags": ["jobs"],
        "summary": "Jobs List",
        "operationId": "jobs_list_api_jobs_get",
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "", "title": "Date From" }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "", "title": "Date To" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "integer" }, { "type": "null" }],
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Offset" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/JobListOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}": {
      "get": {
        "tags": ["jobs"],
        "summary": "Job Detail",
        "operationId": "job_detail_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Job Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/JobDetailOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/documents/batch": {
      "post": {
        "tags": ["jobs"],
        "summary": "Documents Batch",
        "operationId": "documents_batch_api_documents_batch_post",
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/BatchIn" } }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BatchAcceptedOut" } }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{job_id}/relaunch": {
      "post": {
        "tags": ["jobs"],
        "summary": "Job Relaunch",
        "description": "Relance les documents en erreur d'un job (nouveau lot).",
        "operationId": "job_relaunch_api_jobs__job_id__relaunch_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "title": "Job Id" }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BatchAcceptedOut" } }
            }
          },
          "400": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchAcceptedOut": {
        "properties": {
          "job_id": { "type": "string", "title": "Job Id" },
          "repo_job_id": { "type": "integer", "title": "Repo Job Id" }
        },
        "type": "object",
        "required": ["job_id", "repo_job_id"],
        "title": "BatchAcceptedOut"
      },
      "BatchIn": {
        "properties": {
          "kind": { "type": "string", "title": "Kind" },
          "document_ids": {
            "items": { "type": "integer" },
            "type": "array",
            "title": "Document Ids"
          },
          "mailjet_template_id": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Mailjet Template Id"
          }
        },
        "type": "object",
        "required": ["kind", "document_ids"],
        "title": "BatchIn"
      },
...
      "ErrorBody": {
        "properties": {
          "code": { "type": "string", "title": "Code" },
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["code", "message"],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": { "error": { "$ref": "#/components/schemas/ErrorBody" } },
        "type": "object",
        "required": ["error"],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobDetailOut": {
        "properties": {
          "job": { "$ref": "#/components/schemas/JobOut" },
          "items": {
            "items": { "$ref": "#/components/schemas/JobItemOut" },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["job", "items"],
        "title": "JobDetailOut"
      },
      "JobItemOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "patient_id": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Patient Id"
          },
          "patient_display": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Patient Display"
          },
          "document_id": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Document Id"
          },
          "statut": { "type": "string", "title": "Statut" },
          "message": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Message" },
          "has_file": { "type": "boolean", "title": "Has File", "default": false }
        },
        "type": "object",
        "required": ["id", "statut"],
        "title": "JobItemOut"
      },
      "JobListOut": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/JobOut" },
            "type": "array",
            "title": "Items"
          },
          "total": { "type": "integer", "title": "Total" }
        },
        "type": "object",
        "required": ["items", "total"],
        "title": "JobListOut"
      },
      "JobOut": {
        "properties": {
          "id": { "type": "integer", "title": "Id" },
          "kind": { "type": "string", "title": "Kind" },
          "doc_type": { "type": "string", "title": "Doc Type" },
          "statut": { "type": "string", "title": "Statut" },
          "total": { "type": "integer", "title": "Total" },
          "done": { "type": "integer", "title": "Done" },
          "ok": { "type": "integer", "title": "Ok" },
          "skipped": { "type": "integer", "title": "Skipped" },
          "errors": { "type": "integer", "title": "Errors" },
          "created_at": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "doc_type",
          "statut",
          "total",
          "done",
          "ok",
          "skipped",
          "errors"
        ],
        "title": "JobOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}